requirements.txt
Django
django-filter
psycopg2-binary
requests
celery
redis
djangorestframework>=3.12
markdown>=3.3.3
djangorestframework-simplejwt
PyJWT
gunicorn
